requirements.txt
ccxt>=4.0.0
pyyaml>=6.0
aiohttp>=3.9.0
websockets>=12.0
pandas>=2.0.0
